Refactor AboutModal
Refactor BackupRedeemScriptModal

Refactor AddMixerAccountsModal

Refactor AutobuyerRunningModal
