+++ app/components/modals/BackupRedeemScriptModal.jsx
@@ -1,16 +1,18 @@
-import InfoConfirmModal from "../InfoConfirmModal";
+import InfoConfirmModal from "./InfoConfirmModal";
 import { FormattedMessage as T } from "react-intl";
+import { classNames } from "pi-ui";
+import style from "./Modals.module.css";
 
 const Content = ({ script, key }) => (
   <>
-    <div key={key} className="backup-redeem-script-message">
+    <div key={key} className={style.backupRedeemScriptModalMessage}>
       <T
         id="backup.redeem.script.message"
         m={`You must make a backup of the redeem script.
           If the VSP closes without notifying and you do not have access to your redeem script {warnMessage}`}
         values={{
           warnMessage: (
-            <span className="error bold">
+            <span className={classNames("error","bold")}>
               <T
                 id="backup.redeem.script.warn"
                 m="your funds may be lost forever."
@@ -20,13 +22,13 @@ const Content = ({ script, key }) => (
         }}
       />
     </div>
-    <div className="backup-redeem-script-title">
+    <div className={style.backupRedeemScriptModalTitle}>
       <T id="backup.redeem.script.title" m="Script" />
     </div>
     <textarea
       disabled
       value={script}
-      className="backup-redeem-script-textarea"
+      className={style.backupRedeemScriptModalTextarea}
     />
   </>
 );
