+++ app/components/modals/index.js
@@ -7,6 +7,6 @@ export { default as InfoDocumentModal } from "./InfoDocumentModal";
 export { default as PassphraseModal } from "./PassphraseModal/PassphraseModal";
 export { default as ConfirmModal } from "./ConfirmModal";
 export { default as SeedCopyConfirmModal } from "./SeedCopyConfirmModal/SeedCopyConfirmModal";
-export { default as AboutModal } from "./AboutModal/AboutModal";
+export { default as AboutModal } from "./AboutModal";
 export { default as BackupRedeemScriptModal } from "./BackupRedeemScriptModal";
 export { default as AddMixerAccountsModal } from "./AddMixerAccountsModal";
